Move light-specific helpers into `light.helpers`
